Начало реализации функции запросов на добавление новых вещей (не закоммитил своевременно)
