--- src/main/java/ru/practicum/shareit/request/ItemRequest.java
@@ -1,17 +1,0 @@
-package ru.practicum.shareit.request;
-
-import lombok.Data;
-import ru.practicum.shareit.user.model.User;
-
-import java.time.Instant;
-
-/**
- * TODO Sprint add-item-requests.
- */
-@Data
-public class ItemRequest {
-    private Long id;
-    private String description;
-    private User requestor;
-    private Instant created;
-}
