--- src/main/java/ru/practicum/shareit/request/Request.java
@@ -1,22 +1,0 @@
-package ru.practicum.shareit.request;
-
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.Setter;
-import ru.practicum.shareit.user.model.User;
-
-@Getter
-@Setter
-@Entity
-@Table(name = "REQUESTS")
-public class Request {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "ID", nullable = false)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "REQUESTOR_ID")
-    private User requestor;
-
-}
