+++ src/main/java/ru/practicum/shareit/request/dto/ItemRequestWithResponsesDto.java
@@ -3,11 +3,13 @@ package ru.practicum.shareit.request.dto;
 import lombok.Data;
 
 import java.time.Instant;
+import java.util.List;
 
 @Data
-public class ItemRequestDto {
+public class ItemRequestWithResponsesDto {
     private Long id;
     private String description;
-    private Long requestorId;
     private Instant created;
+    private List<ItemResponseShortDto> items;
 }
+
