+++ src/main/java/ru/practicum/shareit/request/dto/ItemResponseShortDto.java
@@ -1,0 +1,11 @@
+package ru.practicum.shareit.request.dto;
+
+import lombok.Data;
+
+@Data
+public class ItemResponseShortDto {
+    private Long itemId;
+    private String name;
+    private Long ownerId;
+}
+
