+++ src/main/java/ru/practicum/shareit/request/model/ItemRequest.java
@@ -1,0 +1,31 @@
+package ru.practicum.shareit.request.model;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import ru.practicum.shareit.user.model.User;
+
+import java.time.Instant;
+
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@Table(name = "item_requests")
+public class ItemRequest {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String description;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "requestor_id")
+    private User requestor;
+
+    @Column(nullable = false)
+    private Instant created;
+}
