+++ src/main/resources/schema.sql
@@ -6,10 +6,25 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 -- Таблица запросов на добавление новых вещей
-CREATE TABLE IF NOT EXISTS requests (
+CREATE TABLE IF NOT EXISTS item_requests
+(
     id BIGINT GENERATED BY DEFAULT AS IDENTITY PRIMARY KEY,
-    description VARCHAR(1000) NOT NULL,
-    requestor_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE
+    description
+    TEXT
+    NOT
+    NULL,
+    requestor_id
+    BIGINT
+    NOT
+    NULL
+    REFERENCES
+    users
+(
+    id
+) ON DELETE CASCADE,
+    created TIMESTAMP NOT NULL DEFAULT NOW
+(
+)
 );
 
 -- Таблица вещей
@@ -19,7 +34,11 @@ CREATE TABLE IF NOT EXISTS items (
     name VARCHAR(255) NOT NULL,
     description TEXT NOT NULL,
     is_available BOOLEAN NOT NULL,
-    request_id BIGINT REFERENCES requests (id) ON DELETE SET NULL
+    item_request_id BIGINT REFERENCES item_requests
+(
+    id
+)
+                                                  ON DELETE SET NULL
 );
 
 -- Таблица бронирования вещей
@@ -38,5 +57,7 @@ CREATE TABLE IF NOT EXISTS comments (
     text TEXT NOT NULL,
     item_id BIGINT NOT NULL REFERENCES items(id) ON DELETE CASCADE,
     author_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
-    created TIMESTAMP NOT NULL
+    created TIMESTAMP NOT NULL DEFAULT NOW
+(
+)
 );
